requirements.txt
python-rapidjson>=0.2.5
djangorestframework>=3.5.4
